requirements.txt
langchain>=0.1.0
langchain-community>=0.0.10
langchain-core>=0.1.10
langchain-openai>=0.0.2
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.25
alembic>=1.13.1
fastapi>=0.109.2
uvicorn>=0.27.1
python-dotenv>=1.0.0
openai>=1.12.0
pydantic>=2.5.2
tiktoken>=0.5.1
gunicorn>=21.2.0
python-telegram-bot>=20.7 
langgraph>=0.3.29
langchain_anthropic>=0.3.10
telegramify-markdown>=0.5.1